Combined notification pop-up and modal
